// rootpilot-frontend-v1-archived/src/types/backend.ts
export interface Incident {
  id: number;
  serviceName: string;
  endpoint: string;
  statusCode: number;
  latency: number;
  exceptionType: string;
  version: string;
  timestamp: string;
}